refactor: Pull error struct into own module
